chore: refactor code into component files + unDRY
- This PR mainly refactors the App.js file into various components.
- I also attempted to reduce the repeating code in miss(), hit(), etc as
  per #4
- In addition, I made minor UI fixes on the coloring of the logs (see
  attached screenshot)

closes #4, #22
