+++ README.md
@@ -1,1 +1,16 @@
 This is an attempt to create a log for killer.
+
+### How to Setup and Contribute
+
+- For the repo
+- `git clone` your fork
+- Add this repo as a remote to make rebasing easier: `git remote add kanyuga https://github.com/kanyuga/killer-counter.git`
+- `git checkout -b <branch-name>`
+- `npm install`
+- `npm start` to start the dev server
+- Work on your feature/task
+- Commit
+- Pull latest `master` from the original repo, e.g. `git pull kanyuga master`
+- Rebase your `<branch-name>` with `master`: `git rebase master`
+- Push to your fork: `git push origin <branch-name>`
+- Create a PR to `@kanyuga/killer-counter`
