+++ src/App.js
@@ -1,139 +1,13 @@
 import React, {Component} from 'react';
-import PropTypes from 'prop-types';
+
+import * as _ from 'lodash';
+
+import BallGrid from './components/BallGrid';
+import Button from './components/Button';
+import PlayerForm from './components/PlayerForm';
+import PlayerList from './components/PlayerList';
+
 import './App.css';
-import * as _ from "lodash";
-
-
-export const Button = (props) => {
-    let classNames = ['btn'];
-    if (props.block) {
-        classNames.push('btn-block');
-    }
-    if (props.context) {
-        let context = props.context;
-        if (props.outline) {
-            context = 'outline-' + props.context;
-        }
-        classNames.push('btn-' + context);
-    }
-    if (props.size) {
-        classNames.push('btn-' + props.size);
-    }
-
-    return <button onClick={props.onClick} className={classNames.join(' ')}> {props.title} </button>;
-};
-
-Button.propTypes = {
-    block: PropTypes.bool,
-    outline: PropTypes.bool,
-    context: PropTypes.oneOf(['primary', 'secondary', 'success', 'info', 'warning', 'link', 'danger']),
-    size: PropTypes.oneOf(['lg', 'sm']),
-    title: PropTypes.string.isRequired
-};
-
-export class PlayerForm extends Component {
-    constructor (props) {
-        super(props);
-        this.state = {
-            name: ''
-        };
-    }
-
-    handleChange = (e) => {
-        this.setState({ name: e.target.value });
-    };
-
-
-    handleSubmit = (e) => {
-        e.preventDefault();
-        this.props.onSubmit(this.state.name);
-        this.setState({ name: '' });
-    };
-
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <div className="input-group">
-                    <input className="form-control"
-                           id="player_name"
-                           name="player_name"
-                           placeholder="Enter Name"
-                           required
-                           autoFocus
-                           value={this.state.name}
-                           onChange={this.handleChange}
-                    />
-                    <span className="input-group-btn">
-                        <button className="btn btn-primary" type="submit">
-                            Add Player
-                        </button>
-                    </span>
-                </div>
-            </form>
-        );
-    }
-}
-
-PlayerForm.propTypes = {
-    onSubmit: PropTypes.func.isRequired
-};
-
-export const BallGrid = (props) => {
-    let buttons = [];
-
-    const className = props.legal ? 'btn-outline-success' : 'btn-outline-warning';
-
-    _.forOwn(props.balls, (ball, number) =>
-        buttons.push(<div key={number} className="col-4 text-center">
-            <button
-                className={'btn btn-ball ' + (ball.active ? className : 'btn-disabled')}
-                disabled={ !ball.active }
-                onClick={() => { props.onClick(number)}} >
-                {number}
-            </button>
-        </div>)
-    );
-
-    //push 1 & 2 to the end
-
-    buttons = buttons.slice(2).concat(buttons.slice(0, 2));
-
-    return (
-        <div className="row">
-            {buttons}
-        </div>
-    );
-};
-
-BallGrid.propTypes = {
-    onClick: PropTypes.func.isRequired,
-    balls: PropTypes.shape().isRequired
-
-};
-
-
-export const PlayerList = (props) => {
-    return (
-        <div>
-            <h3>Score Card</h3>
-            <table className="table table-inverse">
-                <tbody>
-                {props.players.map((player, i) => {
-                    let className = !player.active
-                        ? 'text-muted eliminated-player'
-                        : (i === props.currentPlayer ? 'bg-info current-player' : '');
-
-                    return <tr key={i} className={className}>
-                        <td>{player.name}:</td>
-                        <td className="text-right">{player.points}</td>
-                    </tr>
-                })}
-                </tbody>
-            </table>
-        </div>
-    );
-};
-
 class App extends Component {
 
     ballCount = 15;
@@ -235,65 +109,76 @@ class App extends Component {
         this.setState({ballGridActive: true, ballGridLegal: legalPort});
     };
 
-    hit = () => {
-        const current_player = this.state.players[this.state.currentPlayer];
-        this.setGameState({
+    play = (type, number=this.state.currentBall) => {
+        const typeLog = {
+            hit: 'hit',
+            foulHit: 'foul hit',
+            port: 'ported',
+            miss: 'missed',
+        };
+        let players = this.state.players.slice();
+        let currentPlayer = this.state.players[this.state.currentPlayer];
+        const balls = Object.assign({}, this.state.balls);
+
+        const state = {
             currentPlayer: this.getNextPlayer(this.state.players),
             ballGridActive: false,
-            playLog: this.addLogEntry(current_player.name + ' hit ' + this.state.currentBall)
-        });
-    };
+            playLog: this.addLogEntry(`${currentPlayer.name} ${typeLog[type]} ${number || this.state.currentBall}`, type),
+        };
+        
+        switch(type) {
+            case 'foulHit':
+                currentPlayer = players[this.state.currentPlayer];
+                if (this.ballHasBeenPorted(this.state.balls)) {
+                    const foulBall = this.state.balls[number];
+                    currentPlayer.points -= foulBall.points;
+                    players = this.refreshActivePlayers(players, this.state.balls);
+                }
+                // update state
+                Object.assign(state, {
+                    players,
+                    currentPlayer: this.getNextPlayer(players),
+                    ballGridActive: false,
+                });
+                break;
+            
+            case 'port':
+                const portedBall = balls[number];
+                currentPlayer = players[this.state.currentPlayer];
+                currentPlayer.points += portedBall.points;
+                portedBall.active = false;
+                players = this.refreshActivePlayers(players, balls);
 
-    foulHit = (number) => {
-        let players = this.state.players.slice();
-        const currentPlayer = players[this.state.currentPlayer];
-        if (this.ballHasBeenPorted(this.state.balls)) {
-            const foulBall = this.state.balls[number];
-            currentPlayer.points -= foulBall.points;
-            players = this.refreshActivePlayers(players, this.state.balls);
+                delete state.currentPlayer; // unchanged
+                Object.assign(state, {
+                    players,
+                    balls,
+                    currentBall: this.getNextBall(balls),
+                    ballGridActive: false,
+                });
+                break;
+            
+            case 'miss':
+                currentPlayer = players[this.state.currentPlayer];
+                if (this.ballHasBeenPorted(balls)) {
+                    const missedBall = balls[this.state.currentBall];
+                    currentPlayer.points -= missedBall.points;
+                    players = this.refreshActivePlayers(players, balls);
+                }
+
+                Object.assign(state, {
+                    players,
+                    balls,
+                    currentPlayer: this.getNextPlayer(players),
+                    ballGridActive: false,
+                });
+                break;
+
+            default:
+                break;
         }
-        this.setGameState({
-            players: players,
-            currentPlayer: this.getNextPlayer(players),
-            ballGridActive: false,
-            playLog: this.addLogEntry(currentPlayer.name + ' foul hit ' + number)
-        })
-    };
-
-    port = (number) => {
-        let players = this.state.players.slice();
-        const balls = Object.assign({}, this.state.balls);
-        const currentPlayer = players[this.state.currentPlayer];
-        const portedBall = balls[number];
-        currentPlayer.points += portedBall.points;
-        portedBall.active = false;
-        players = this.refreshActivePlayers(players, balls);
-        this.setGameState({
-            players: players,
-            balls: balls,
-            currentBall: this.getNextBall(balls),
-            ballGridActive: false,
-            playLog: this.addLogEntry(currentPlayer.name + ' ported ' + number)
-        });
-    };
-
-    miss = () => {
-        let players = this.state.players.slice();
-        const balls = Object.assign({}, this.state.balls);
-        const currentPlayer = players[this.state.currentPlayer];
-
-        if (this.ballHasBeenPorted(balls)) {
-            const missedBall = balls[this.state.currentBall];
-            currentPlayer.points -= missedBall.points;
-            players = this.refreshActivePlayers(players, balls);
-        }
-        this.setGameState({
-            players: players,
-            balls: balls,
-            currentPlayer: this.getNextPlayer(players),
-            ballGridActive: false,
-            playLog: this.addLogEntry(currentPlayer.name + ' missed ' + this.state.currentBall)
-        });
+        
+        this.setGameState(state);
     };
 
     ballHasBeenPorted = (balls) => {
@@ -323,9 +208,9 @@ class App extends Component {
         });
     };
 
-    addLogEntry = (logEntry) => {
+    addLogEntry = (log, type='none') => {
         const playLog = this.state.playLog.slice();
-        playLog.push(logEntry);
+        playLog.push({ log, type });
         return playLog;
     };
 
@@ -352,24 +237,22 @@ class App extends Component {
             } else {
                 display = (
                     <div className="action-section">
-                        <h3>Current Player: {this.state.players[this.state.currentPlayer].name}</h3>
+                        <h3>Current Player: <span className="current-player">{this.state.players[this.state.currentPlayer].name}</span></h3>
                         <h3>Current Ball: {this.state.currentBall}</h3>
 
                         <Button onClick={() => this.showBallGrid(true)} size="lg" block context="success" title="Port"/>
                         {this.state.ballGridActive && this.state.ballGridLegal ?
-                            <BallGrid balls={this.state.balls} legal={true} onClick={this.port}/> :
+                            <BallGrid balls={this.state.balls} legal={true} onClick={(number) => this.play('port', number)}/> :
                             null
                         }
-                        <Button onClick={this.hit} size="lg" block context="info" title="Hit" />
-                        <Button onClick={this.miss} size="lg" block context="danger" title="Miss" />
+                        <Button onClick={() => this.play('hit')} size="lg" block context="info" title="Hit" />
+                        <Button onClick={() => this.play('miss')} size="lg" block context="danger" title="Miss" />
                         <Button onClick={() => this.showBallGrid(false)} size="lg" block context="danger" title="Foul Hit" />
                         { this.state.ballGridActive && !this.state.ballGridLegal ?
-                            <BallGrid balls={this.state.balls} legal={false} onClick={this.foulHit}/> :
+                            <BallGrid balls={this.state.balls} legal={false} onClick={(number) => this.play('foulHit', number)}/> :
                             null
                         }
-
                     </div>
-
                 );
             }
         }
@@ -400,8 +283,8 @@ class App extends Component {
                     </div>
                     <div className="col-6 col-sm-3">
                         <h3>Log</h3>
-                        <ol>
-                            {this.state.playLog.map((playLog, i) => <li key={i}>{ playLog }</li>)}
+                        <ol className="play-log">
+                            {this.state.playLog.map((playLog, i) => <li className={playLog.type} key={i}>{ playLog.log }</li>)}
                         </ol>
                         { this.state.history.length > 1
                             ? <Button onClick = { this.undo } block outline context="warning" title='Undo'/>
